//maddy

+++ src/main/java/com/ats/webapi/model/SpecialCake.java
@@ -92,23 +92,23 @@ public class SpecialCake implements Serializable {
 	
 	
 	@Column(name="mrp_rate1")
-	private int mrpRate1;
+	private float mrpRate1;
 	
 	@Column(name="mrp_rate2")
-	private int mrpRate2;
+	private float mrpRate2;
 	
 	@Column(name="mrp_rate3")
-	private int mrpRate3;
+	private float mrpRate3;
 	
 	@Column(name="sp_rate1")
-	private int spRate1;
+	private float spRate1;
 	
 	@Column(name="sp_rate2")
-	private int spRate2;
+	private float spRate2;
 	
 	
 	@Column(name="sp_rate3")
-	private int spRate3;
+	private float spRate3;
 	
 	@Column(name="is_slot_used")
 	private int isSlotUsed;
@@ -155,43 +155,42 @@ public class SpecialCake implements Serializable {
 	public void setIsAddonRateAppli(int isAddonRateAppli) {
 		this.isAddonRateAppli = isAddonRateAppli;
 	}
-	public int getMrpRate1() {
+	public float getMrpRate1() {
 		return mrpRate1;
 	}
-	public void setMrpRate1(int mrpRate1) {
+	public void setMrpRate1(float mrpRate1) {
 		this.mrpRate1 = mrpRate1;
-	}
-	public int getMrpRate2() {
+	}	
+	public float getMrpRate2() {
 		return mrpRate2;
 	}
-	public void setMrpRate2(int mrpRate2) {
+	public void setMrpRate2(float mrpRate2) {
 		this.mrpRate2 = mrpRate2;
 	}
-	public int getMrpRate3() {
+	public float getMrpRate3() {
 		return mrpRate3;
 	}
-	public void setMrpRate3(int mrpRate3) {
+	public void setMrpRate3(float mrpRate3) {
 		this.mrpRate3 = mrpRate3;
 	}
-	public int getSpRate1() {
+	public float getSpRate1() {
 		return spRate1;
 	}
-	public void setSpRate1(int spRate1) {
+	public void setSpRate1(float spRate1) {
 		this.spRate1 = spRate1;
 	}
-	public int getSpRate2() {
+	public float getSpRate2() {
 		return spRate2;
 	}
-	public void setSpRate2(int spRate2) {
+	public void setSpRate2(float spRate2) {
 		this.spRate2 = spRate2;
 	}
-	public int getSpRate3() {
+	public float getSpRate3() {
 		return spRate3;
 	}
-	public void setSpRate3(int spRate3) {
+	public void setSpRate3(float spRate3) {
 		this.spRate3 = spRate3;
 	}
-	
 	public int getSpId() {
 		return spId;
 	}
